requirements.txt
torch
torchvision
pandas
Pillow
matplotlib
flwr
